Bug#31582622 DOWNGRADE TESTS FAILING BECAUSE OF ARBITRATION ERROR
Modifies testUpgrade.cpp to add 5 seconds of sleep after management
nodes are restarted.

TESTDOWNGRADE -N DOWNGRADE_NR1 and -N DOWNGRADE_API_BEFORE_NR1 are
failing because the data nodes aren't able to find their arbitrator
since the arbitration setup isn't complete when both the management
nodes are stopped first and restarted one after the other. So, before
data nodes are attempted to restart, a delay of 5 seconds is added to
give management nodes to complete the arbitration setup.

This "sleep" statement along with a couple of more similar statements
already existing will be replaced with a stronger solution as part of
the Bug#31648682 using the analysis from Bug#31582622.

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>