Bug#31096309 - 8.0.16+ DIFFERENT RESULTSET THAN 5.7 AND <=8.0.15
Cause:

A semijoin strategy is chosen for the join of a correlated subquery
which has a OFFSET clause that indicates that the subquery must contain
at least two rows. This will result in ignoring the OFFSET clause and
return invalid rows.

Fix:

Do not select semjoin strategy when a OFFSET clause requires more than
one row or select limit is 0.

Change-Id: Ib3752d7b5191b51afbe2dbfeaa0e8436b67c0751