Bug #31471959 NDB_SUMA_HANDOVER FAILS WITH MGMD CRASH
The mtr test ndb_suma_handover fails because the data node returns
node info with a zero ndb-version and a non-zero mysql-version. This
causes a crash since ndb_mgmd requires that either both should be
zero or both should be nonzero. The version mismatch arises from
QMGR, which has multiple cases where the ndb-version is zeroed without
modifying the mysql_version. Also, the API_VERSION_REQ signal handler
has 2 separate calls to get the node info for the ndb-version and the
mysql-version, which can also result in inconsistency if the node
info changes between getNodeInfo() calls.

This is fixed by updating QMGR to modify the mysql_version whenever
the ndb version is modified. In addition, the API_VERSION_REQ handler
is changed so that the ndb-version and mysql-version are set from a
single NodeInfo version.

Change-Id: Ic45d13b075df6129da35980c6b07e7a094f59c36