Bug#31686096 - X.CONNECTION_UNIXSOCKET FAILS REALY AT SOCKET-FILE CLEANUP
Description
===========
Race-condition between thread that initializes `X Plugin` and a thread
that shutdowns the server. The second thread might mark the `listener`
object that it freed its resources before they are full initialized.

Fix
===
The interaction between `Socket_acceptor_task` and `Listener_unixsocket`
was fixed in a way, that the first alaways request to clean resources
on a listener.

RB: 24915
Reviewed by: Grzegorz Szwarc <grzegorz.szwarc@oracle.com>