Bug #31594774 : NDB.TEST_MGMD FAILS ON PB2
Problem
-------

In testMgmd, in runTestBug12352191, both the mgmds are configured with HostName=localhost.
Without WL#12963 (introduces IPv6 in NDB), localhost resolves to 127.0.0.1.
However, after WL#12963, depending on the entries in /etc/hosts, an mgmd sees
itself and the other one's address as either ::1 or ::ffff:127.0.0.1.
This causes the check which expects both the mgmds' addresses to be 127.0.0.1
to fail.

Solution
--------

Make necessary changes so that IPv4 numeric addresses are returned without
the IPv4-mapped-IPv6 address prefix.
Also revise the check in runTestBug12352191 to expect the addresses to either
be ::1 or 127.0.0.1

Patch also fixes another bug found during code inspection where AF_INET was used
instead of AF_INET6

Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>