Bug#31096384: SELECT .. LOCK IN SHARE MODE REQUIRES LOCK  TABLES GRANT
When SKIP LOCKED and NOWAIT were implemented (a.k.a. wl
3597) an additional requirement was added that a user must
have the SELECT select privilege and at least one of UPDATE,
DELETE or LOCK TABLES for both lock strengths SHARE and
UPDATE.

This patch relaxes the requirement for the FOR SHARE lock
strength so that a SELECT privilege is sufficient. FOR
UPDATE is unaffected.

Rationale: Additional privilege requirement caused
compatibility issues for users. OTOH a similar effect as a
FOR SHARE clause can be achieved by users having only SELECT
privileges by using serializable isolation mode. Hence it
doesn't make much sense to require additional privileges for
the FOR SHARE clause specifically.

Change-Id: Ic96fd3f2beaaed3bc19fd077c8e90f5582559061