WL#12532: Make number of main and rep threads configurable
Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>

This worklog makes it possible to create configuration with
main and rep threads combined. This is useful in smaller
setups where e.g. we want to use 1 main+rep thread, 1 ldm thread
and 1 recv thread.

In addition it creates a single-threaded ndbmtd configuration
where the recv thread is also main+rep+ldm+tc. This makes it
possible to deprecate ndbd such that everyone can use ndbmtd
and a lot of ndbd code can be removed.

When selecting ndbinfo table threadblocks it was possible with
new config where we have main and rep threads in the same
thread as ldm thread we could overrun the buffer used when
calling mt_get_blocklist. Modified the constant
MAX_INSTANCES_PER_THREAD to be more correct.