Bug #31655906 ACQUIRE_RELATED SHOULD RETURN ERROR IF RELATED SERVICE IS NOT FOUND
Description:
-----------
acquire_related() takes:
- requested service name
- handle of already acquired service
- handle of registry

and tries to find requested service that is related to already acquired
service.

If it fails it falls back to default implementation.

This may create problem because caller may be expecting that since
services
are related, they can operate upon same shared object.

Fix:
---
If the related service is not found, returning the 'true' instead of
default service.

RB#24945