WL#14107: Introduce PartitionsPerNode
Approved-by: Sanjana DS <sanjana.ds@oracle.com>

This worklog introduces a new configuration parameter called
PartitionsPerNode. This parameter steers the number of partitions
created by a default table. By default this would be equal to the
number of LDM threads in the node. But in this case we will instead make
it possible to have the number of default partitions be smaller
such that a table don't make use of all LDM threads. This introduces
a small unbalanced behaviour, but makes it possible to avoid
splitting up the tables into too many parts when the number of
LDMs grows high. This prepares to make it possible to increase the
number of log parts and LDMs to 64 when introducing query threads.

By default it is set to 0 which means that the default number of
partitions per node is equal to the number of LDM threads in the
nodes.