BUG#31726568: findTransaction issues
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

findTransaction in DBLQH cannot handle the test case testOperations where we can use
the same operation record, the same transaction id, but from different TC nodes.
Improved uniqueness of operations in transaction hash in DBLQH.

The fix for findTransaction assumed that the block reference to the failing
DBTC instance was correct, or at least pointing to the right node and
right block instance. This wasn't correct, the variable was not
initialised.

Thus we fall back to not checking the block reference at all in this case.
Failing to handle this properly led to locked rows remaining after a node
failure thus leading to node recovery not being able to complete.

Prepared findTransaction for a version where the block reference is properly
filled in.