WL #13459: Optimize hash table in hash join [patch 6/9, not multimap]
Use an std::unordered_map instead of std::unordered_multimap. The motivating
reason for this is that we're going to change to a third-party hash table in a
later patch, and most third-party implementation don't support multimaps.
However, it also allows us to make for more efficient encoding of nodes that
share the same key. Instead of inserting multiple entries into the map
(usually including multiple copies of the key), we can link them together into
a single-linked list. Since we know that the pointers are likely to be very
close together (they come from the same MEM_ROOT), we can store the difference
between them as a Varint128, which takes down the pointer size from 8 to
(typically) one or two bytes. This is encapsulated in a new class called
LinkedImmutableString.

Note that we don't store the length of the value anymore; we never used it,
as pack()/unpack() keeps track of lengths itself.

There's also a similar class called ImmutableStringWithLength. We don't
use it yet, but soon, it will be used for storing keys in a similar fashion
(which don't have a next pointer, but do need the string length).

We include some Varint128 parsing routines pretty much verbatim from protobuf,
since version 3.6 (which we still need to support) doesn't expose these
to clients.

Change-Id: I464ea2b4af1d73ba9e78a5a623b8ec29de66439e