Bug #31180508   NDB : BACKUP NOT O_DIRECT SIZE COMPLIANT IN SOME CASES
Backups error out with FsErrInvalidParameters when the filesystem is
running with O_DIRECT and a datafile write is not aligned with the
512-byte block size used by O_DIRECT writes. If the total fragment
size in the datafile is not aligned with the O_DIRECT block size,
the last write is padded to the required size. However, if there are
no fragments to write to a datafile, BACKUP writes only the header
and footer to the datafile. Since the header and footer are less than
512 bytes, this causes the O_DIRECT write to fail.

This is fixed at the time of closing the datafile, by padding out
the generic footer to the 512-byte size using an EMPTY_ENTRY.

Self-checking is added to NdbFS on async reads/writes, ensuring that all
O_DIRECT reads and writes are sized and aligned as required. This is
always-on, and independent of whether compression is in use, or the
underlying FS actually supports O_DIRECT.

The default data node config for the ndb testsuite is modified to
enable O_DIRECT.

Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>
Change-Id: I11c9b38f5418211c0c519bf600fdf31376705b7f