Bug#31396947 PAGE COMPRESSION ISN'T ENABLED AFTER IMPORT
Issue
=====
As part of IMPORT we read all the pages in the tablespace and write it back to
the disk, after updating the page header with relevant information like the
space id.

At this point we do not have a valid fil_space_t object associated with the
tablespace which is where the compression information is stored. So at the time
of write the IORequest does not know to compress the page as the information is
not available, thereby writing the page to the disk as an uncompressed page.

So, after the table import of a compressed table the pages of the imported
table are not compressed even though information in INFORMATION_SCHEMA suggests
otherwise.

Fix
===
Store the compression information in dict_table_t object as well and at the
time of page write mentioned above pass the information to IORequest thereby
allowing the page to be compressed.

Consequently set the compression information in fil_space_t onject once it's
created so that the compression is carried out after the table is imported.

RB: 24660
Reviewed-by: mateusz.rukowicz@oracle.com
Reviewed-by: deepa.dixit@oracle.com