Bug#31742798 - GROUP REPLICATION NOTIFICATION CRASHES X PLUGIN WHEN THERE IS NOT ACCEPTED CLIENT
Description
===========
Race-confition between thread that starts handling new connection
and a thread that 'dispatches' group-replication notifications.
There is a short period when `xpl::Client` has session not
set. The problem is that dispatcher requires the session.

Fix
===
Group-replication notice dispatcher thread must check if
the session pointer is valid.

RB: 24969
Reviewed by: Grzegorz Szwarc <grzegorz.szwarc@oracle.com>