Bug#31603289: REGRESSION: CRASH IN TABLE_LIST::GET_CLONE_FOR_DERIVED_EXPR, ASSERTION `FALSE' IN               TABLE_LIST::GET_DERIVED_EXPR
Problem:
While pushing conditions down to derived table, columns in a condition are
replaced with derived table expressions. To get the derived table expression
corresponding to the column being replaced, we use the field index. However
during query transformations, there is a possibility that new conditions
are added with multiple references to an existing condition. When such a
condition is pushed down, after the first reference to a column is replaced
with derived table expression, when the replacement for second
reference to the same column is being done, the field index is no more
valid because the column has already been replaced with derived table
expression. This results in server exit.

Solution:
Check that the table reference of a column that is getting replaced is
the same as the derived table from where the replacements are fetched.
Replacement is necessary only in such a case.

Approved By: Gleb Shchepa <gleb.shchepa@oracle.com>