Bug#31280526: 8.0.20+ CRASH IN JSON_BINARY::PARSE_BINARY
Problem:
A dynamic range scan will run range optimizer for every row
fetched from the first table in a join to see if a range scan
can be picked for the second table using the value available
from first table's row. If there are no right indexes, a table
scan could be picked too. W.r.t the query in bugpage,
it picks a table scan once followed by a range scan on a non-covering
index. The dynamic range iterator has two read sets which
are used for both these cases. One read set, used for table scan
which includes the base columns of generated columns required
for processing the query. And another read set which does
not include the base columns in the read set used for range scans.
The reason for this is, in case of covering indexes, read set
should not include base columns to avoid adding unneeded columns
by hash join/BKA.
However the same read set is used for a non covering index too.
This results in an assert later as the bits are expected to be set.

Solution:
When initializing table read set in dynamic range iterator, we now
set it to include base columns when range optimizer picks a
non-covering index.

Change-Id: I67e4dd3714e9d529a0b5b9b4ad0a37ce55f1e2d7