Bug #31749396: OUT OF THREAD STACK IN CREATEITERATORFROMACCESSPATH
Certain Clang versions have very inefficient stack usage when there is no -O
option and debugging is active, causing deep multi-way joins to run out of
stack in CreateIteratorFromAccessPath. (This is not a problem in production.)
Consistently use the “const auto &param” pattern to reduce the number of
temporaries dramatically, cutting the stack frame usage by about 25% for this
function.

Change-Id: Idd33feeff05d9410ec01e9a144f6a4774dee9045