WL#14059 REMOVES GENERATION OF COVERAGE HTML REPORT
Modifies atrt-compute-coverage.sh script to not generate coverage html
report.

Currently at end of every test suite run, html report was generated
using coverage.info for that run. Though the coverage was calculated
only for ndb source code, html report generated was found to be more
than 100MB and the compressed form was found to be more than 10MB.

For the frequency of test runs, storing such big reports for each test
suite does not seem to be a good solution. In this change the generation
of html reports using genhtml tool has been removed.

Change-Id: I47dba1d1b516b883f9d42b9f73822084aac0612c