Bug#31599885: Global-buffer-overflow in decimal_bin_size_inline
The main problem here is that when aggregating type of a BIT
expression into a variable, we set a confusing type: data_type()
is VARCHAR, whereas the result_type() is INT_RESULT. When trying
to aggregate this misleading information into an AVG operation, we
end up with a DECIMAL type that has inconsistent scale and precision
values, and we end up with an assert.

The solution to the problem is in two parts:

- First, adjust the type_for_variable() function to return
  MYSQL_TYPE_LONGLONG for a BIT column, instead of VARCHAR. This is
  what causes the main confusion, and it is also more reasonable to
  store a BIT column as a numeric value.

- Second, refactor some resolve_type() logic to use set_data_type_xxx()
  functions to enhance consistency in type resolving.
  Also ensure that the result type is always derived from data_type()
  in these functions.
  This applies to classes Item_func_set_user_var, Item_sum_sum and
  Item_sum_avg.

The logic for calculating decimals for aggregate operations for a float
value with specified scale and precision is also made more explicit.

The refactoring also made it possible to remove function
set_data_type_from_result(), which did not set data type consistently.

Reviewed by: Tor Didriksen <tor.didriksen@oracle.com>