Bug#30943963 - PREPARE STATEMENT CRASHED THE SERVER
Cause:

When explicit locks are taken on a view and the same view name used in
the prepare of a create table statement, the server exits. The prepare
statement should pass and the execution will fail as the table name
already exists.

When explicit locks are taken, we try to open and parse the view
definition even if it is a table name for create. This view expansion
and applying transformations on the views' tables while processing a
table for create causes server exit.

Fix:

In the prepare of a statement with lock tables mode, if the table is for
create and if matches a an existing view, skip expansion of the view.

Change-Id: I97a4d005f2ba329bde74578db9c3b890c8c91707