Bug #31790217: REWRITE STREAMING AGGREGATION Bug #22588319: OUTER REFERENCE TO AGGREGATE INCORRECTLY 0 OR NULL IN SUBQUERY [noclose] Bug #27272052: REGRESSION: GOT NULL WHEN ROW() FUNCTION USED WITH IN OPERATOR FOR INNODB TABLE Bug #31073167: WRONG RESULT FOR EXPRESSION INVOLVING GROUP BY COLUMNS WITH ROLLUP Bug #31868610: DATA TYPE CHANGED IN A GROUP BY WITH ROLLUP
Rewrite streaming aggregation, for simplicity and increased modularity.

When aggregating already-sorted data (so-called streaming aggregation,
ie., not using a temporary table), we have a fundamental problem:
We don't know when a group ends until it's too late. When we see the
first row in the next group (or EOF), we know that the current group has
ended, but by then, the group expressions that we would like to output
are usually already overwritten. We don't have a good way of simply
asking the storage engine to keep both rows around (everything points
implicitly into table->record[0]), so we need to deal with this
ourselves.

The current code solves this in two different ways, depending on whether
we are outputting to a temporary table or not. If we are, then
AggregateIterator takes over part of MaterializeIterator's
(or StreamingIterator's) job of copying the data into the destination
table's records. This is fairly complex, especially as group expressions
and aggregate values need to be copied at different times and in
specific internal orders. (It is also a modularity violation. In the
pre-iterator executor, group + HAVING + send was in one combined
function, and group + HAVING + write was in another combined function.
That was even less modular, but the split into iterators also then
spread some of the logic out in unfortunate ways, since
AggregatorIterator ended up taking on some of MaterializeIterator's
duties.) If output goes directly to the client with no temporary table
in-between, we set up a special slice (REF_SLICE_ORDERED_GROUP_BY)
of Item_copy_* items that keep the previous values of the group
expressions. In some places, this is referred to as a “virtual one-row
temporary table”.

Replace all of this by simply saving a representative row for the group
when we encounter it the first time, so that we can easily get its
fields back for the output row. We reuse some infrastructure from hash
join for doing this, and it simplifies a lot of the logic:

 - MaterializeIterator/StreamingIterator now always does the copying job
   (except for windowing functions, which keeps its separate logic
   here). AggregateIterator now never needs to worry about it.

 - PrecomputedAggregateIterator, which existed to copy group expressions
   into a destination table in certain situations, goes away due to the
   previous point. (This may make EXPLAIN FORMAT=tree more confusing,
   as the user may wonder where the group step went, but it's an
   accurate representation of the work we are doing.)

 - All of Item_copy_* goes away. In its place is only a thin shell
   called Item_metadata_copy, which serves a role for keeping track of
   metadata in GREATEST/LEAST type aggregation (Item_copy_* was reused
   for this).

 - The entire REF_SLICE_ORDERED_GROUP_BY slice goes away;
   AggregateIterator now no longer needs to think about slices at all.
   We also don't need setup_copy_fields(), its item-sorting logic,
   and the slice switching in the code calling it.

In the process, we automatically get a fix for most of bug #22588319
(OUTER REFERENCE TO AGGREGATE INCORRECTLY 0 OR NULL IN SUBQUERY);
outer references to aggregates in SELECT lists now work properly,
although they are still broken in ORDER BY. A few such queries in the
RAPID test suite were recorded with wrong answers.

Similarly, this fixes bug #27272052 (REGRESSION: GOT NULL WHEN ROW()
FUNCTION USED WITH IN OPERATOR FOR INNODB TABLE). Again, rapid.row
was recorded with the wrong answer.

It also fixes bug #31073167 (WRONG RESULT FOR EXPRESSION INVOLVING GROUP
BY COLUMNS WITH ROLLUP), since it's no longer tricky to know when to
copy expressions that depend on both ROLLUP NULLs and other group values
(all copying, if any, happens in MaterializeIterator).

There is one new tricky thing around slices. Previously, with rollup,
we'd first set up the base slice, and then start wrapping certain fields
into rollup items (for NULL adding and switching between different
aggregates). This would cause the items in the base slice to no longer
be a match with the field list; however, it would be fine, since rollup
would cause us to always make the REF_SLICE_ORDERED_GROUP_BY slice,
which would use the updated field list as its source. We no longer have
that, so we need to go back and update the base slice(s) after-the-fact.
However, GROUP BY and ORDER BY items could be pointing into that (they
always point into slices, probably for historical reasons), and some of
the rollup processing code depends on _not_ seeing rollup wrapper items
in GROUP BY items. Thus, we have to be careful about updating those
slices only after all the other processing.

The rollup wrappers in HAVING filters are now generally more visible in
EXPLAIN FORMAT=tree, which is probably a good thing.

There is one semi-regression. If you have a nondeterministic expression
in the SELECT list with an alias, and use HAVING against that alias
(a MySQL extension), one might want the expression to only be evaluated
once. WL #12074 caused this to not be true in some cases involving
SELECT DISTINCT, and we decided that this was okay back then. Now,
since we no longer have the Item_copy_* fields snapshotting group
expressions, we get evaluating twice in more situations, notably when
HAVING expressions mix group expressions and aggregates. (The root cause
is that we need to evaluate the item for HAVING, and then don't store
the value we chose when we are to send or materialize the row
immediately afterwards.) We don't have documented behavior around this,
so it should be fine, but it is worth noting. A potential fix could
involve snapshotting the row after GROUP BY and before HAVING, possibly
by means of a StreamingIterator, but it would involve rewriting the
HAVING to point to that snapshot, which we shouldn't do before we have
gotten rid of slices entirely and can do such things without introducing
more of them.

Performance is pretty much neutral; we spend a tiny bit of time saving
and loading the rows (~0.5% on DBT-3 Q16), and win some of it back
by means of reduced complexity in AggregateIterator, in not going through
Item_copy_*, and during setup.

Not accounting for bugs, the AggregateIterator code is now prepared
for dealing with inputs from multiple tables, although the current join
optimizer never sets that up (if there are two or more, it sets up a
temporary table to hold the result of the join). This will be useful for
the hypergraph join optimizer, which generally prefers not to set up
temporary tables unless it confers a direct advantage in some way.

Change-Id: I7297cffde70b83a4042187fe26813301aef9f8ee