BUG#31726514: Problem with testSystemRestart -n Bug54611 T1
Approved-by: Mauritz Sundell <mauritz.sundell@oracle.com>

It is fully possible that DIH receives an LCP to restore of e.g. 12 and then
the RESTORE decides to use LCP 13 to restore this fragment. With the
introduction of Partial LCP DIH isn't fully synchronized with the LCP id
used to restore a fragment. Thus it could be one LCP off the target. This
was checked in an ndbrequire that was specific for the test case:

testSystemRestart -n Bug54611 T1

Fixed the ndbrequire to handle all variants possible now.