BUG#31688797: Reading primary key from hash index can fail
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

When searching the hash index we use the row to read the primary key most
of the time. However if the row hasn't been committed yet we can sometimes
read the primary key from the copy row.

However if the row has been deleted the row can no longer be used to read
the primary key. Previously this meant that we reported the primary key
more or less as a NULL value not equal to anything, but we could also
crash since the comparison was made on uninitialised data.

The extension now is to make comparison only if row hasn't been deleted.
If the row has been deleted we try to find the row among the operations
in the serial queue. If no operation has the primary key, then it is ok
to report any comparison as not equal since no entry in the parallel
queue will reinsert the row.

If an entry in the serial queue is an INSERT then it is vital that we do
report the primary key from this operation, otherwise we could end up
inserting the same primary key twice.