Bug#31582383: QUICK_RANGE_SELECT::INIT_ROR_MERGED_SCAN ALWAYS USES                F_RDLCK
Problem:
Range optimizer does not use correct lock type after it clones
the handler needed to do merged scans. It unconditionally uses
read lock. This results in different side effects for different
scenarios.

For ex:
For a select query with "FOR UPDATE", the requested lock type is a
write lock. However after cloning the handler for an index merge
scan, range optimizer requests a read lock which results in a
mismatch between the original select lock type of the handler and
the requested lock.

Similarly for DD tables because of their special handling, it
would result in setting the lock type to LOCK_NONE.

Solution:
Use the original lock type of the handler even in the cloned
handler.

Reviewed By : Kuba Lopuszanski <jakub.lopuszanski@oracle.com>
Reviewed By : Sergei Glukhov <sergey.glukhov@oracle.com>