Bug#31321518 SPJ BLOCK SHOULD BE MORE EAGER IN RELEASING GLOBALSHAREDMEMORY PAGES
Currrently the SPJ block may be really slow in releasing the shared memory
resorce 'GlobalSharedMemory'. Thus, after a peak usage of such memory objects
in the SPJ block, it will take long (multiple second -> minutes) before
the shared memory has been released to the global pool.

This patch use the existing NdbStatistics class to collect statistics of number
of 'Global Shared' memory pages used by SPJ. That statistics class provide a
rolling average over a number of sampled values, such that older
samples gradually becomes less significant. (Use '5' as the
sample window size in this patch). That provides us with a baseline
for how many globalSharedMemory pages the SPJ block may retain.

Support is added to SPJ to keep track
of number of pages being allocated by SPJ, and number of pages
contained in the RowPage_list / _fifo (use the 'counted' variant of
the lists).

We also keep track of max pages which has been used (m_maxUsedPages)
in each sampling interval. This information is collected into
NdbStatistics object m_usedPagesStat, such that we can 'get' the
mean and standard deviation statistics of max pages being used.

At the sample intervals we check if we currently holds more
allocated memory pages than the upper 95% percentile of the
'maxUsedPages'. If we do, and there are unused pages in the free_list,
any excess pages are release back to the memory manager. A real-time
break is used though to limit number of release pages to 16.

As we would like memory pages to be returned to the global
memory manager faster when there are many pages in the free_list,
the sampling/release intervall are set to 10ms when there are more than
16 free pages, else it is 100ms.

Note that when it is a sudden increase in number of free pages, like
immediately after a large SPJ query has completed, or a SPJ-batch, the
standard deviation will rise abruptly and *increase* the upper95% percentile
for a short while. Thus, in the first sample intervalls after a SPJ
completion no page releases will take place. Such a grace periode
is wanted behaviour, as the completion may only be the completion of
a batch, with a new NEXTREQ needing the freed pages coming shortly
thereafter - or a new SPJ request may arrive.

Reviewed By: Mikael Ronstrom <mikael.ronstrom@oracle.com>