Bug#31774685 GETCOLUMNBYHASH() MAY RETURN AN INCORRECT COLUMN
If the name of one Column was the same as the prefix of another Column
name, these columns could compare as equal in getColumnByHash().

Root cause was that strncmp() was used to compare these, with a length
argument which could be from the shortest of these two colum names.

Patch replace strncmp() with strcmp(), which will compare the full
Column names.

Reviewed By: Frazer Clement <frazer.clement@oracle.com>