WL#12532: Fixes for receive thread as main thread
Approved-by: Sanjana DS <sanjana.ds@oracle.com>

When receive thread is running as main thread and not as the only
thread the freeze used for multi sockets will freeze since the
freeze action is taken by a freeze'd thread.

Also needed to wake up more often when acting as main thread
as receive thread to ensure that CREATE TABLE and other metadata
actions are faster.