Bug #31790217: REWRITE STREAMING AGGREGATION [slices in StreamingIterator, noclose]
Preparatory patch.

StreamingIterator, like MaterializeIterator, needs to contend with the
fact that iterators above it and below it may need to run with different
slices active (specifically, in the cases where it is emulating a temporary
table internal to a JOIN). However, this was not done correctly.

The reason this wasn't caught before now was that in the cases where
we'd actually use StreamingIterator internally in a join and it matters,
AggregateIterator would be there with its own slice setting, and would
happen to set us up in a situation where we got the right result anyway.
This will soon not be the case anymore, so fix it.

Also, the child iterator could be instantiated in the wrong JOIN in some
places -- access paths don't store explicitly which JOIN they belong to
except on transitions between JOINs, so CreateIteratorFromAccessPath() needs
to keep track of which one they are in, and it would forget to do transitions
stemming from StreamingIterator. There are no known actual bugs stemming from
this mistake, but it should be corrected nevertheless.

Change-Id: Ic13150752850418d78231448c379624c3b5cc7b7