Bug#31648682 REMOVE SLEEPS FROM UPGRADE-DOWNGRADE TESTS
Modifies upgrade/downgrade tests to replace 5 seconds sleep after
management nodes are restarted by logic that waits for arbitration to
complete.

The sleep statements were added as a hack to wait for the arbitration
setup to be done for all the data nodes. Arbitration setup is a
multi-step protocol to make sure all the data nodes are registered to a
management node as an arbitrator which is responsible for making the
decision about which part of the cluster should be up when there is a
failure in any one of the nodes. Since the information about the
arbitration status is available in the "membership" table of the
"ndbinfo" database, the sleeps are replaced with a check to make sure
arbitration is setup for all the data nodes before they're attempted to
restart. So, membership table is polled for the arbitration status every
500ms and if the arbitration setup is not complete in 5 seconds,
test fails saying that Arbitration setup took too long.

Approved by: Tiago Alves <tiago.alves@oracle.com>

Change-Id: Ib3915eef74c48382b0b437b6773275abb710c5d1