Bug#31661309: SIG6 IN ADD_KEY_FIELD IN SQL/SQL_OPTIMIZER.CC Bug#31671280: ASSERTION FAILURE `*SARGABLES >               REINTERPRET_CAST<SARGABLE_PARAM *>(*KEY_FIELDS)'
Problem:
During transformation of a grouped query into derived table, the where clause
and the having clause become part of the derived table. However the condition
count is not updated for the derived table. This results in reduced memory
allocation while creating keys for ref access.

Solution:
Update the condition count of the derived table correctly.

Change-Id: Ib8eb95ac69ce8f613a2bbe0b6689a294662b496b