BUG#31795721: ENUMERATION VALUE ENUM_SESSION_TRACK_GTIDS::OFF DEFINED AS A GLOBAL SYMBOL
Background:

The C++ enum type enum_session_track_gtids was defined in global scope
and had a member named 'OFF'.  This resulted in OFF being a global
symbol.

Problem:

The name is too broad and the use case too narrow for it to live in
global scope.  (In fact, it was used by mistake in one place where
another type was expected.)

Fix:
Prefix the members of this enum type with SESSION_TRACK_GTIDS .

RB:25022