Bug#31787736 MTR TEST FAILURE IN RAPID.AUTO_PROVISIONING BECAUSE OF FIX FROM BUG#31564407
Issue
=====
An assert in ha_innobase::init() was added as part of the fix for BUG#31564407
assuming that the trx_t::isolation_level would be up to date with
THD::tx_isolation. But turns out this is not always the case.

Fix
===
On looking closer, it was observed that the InnoDB histogram sampling does not
really need transaction support as we anyway do READ_UNCOMMITED, and the
parallel reader requires trx only to do MVCC which would not be required
in our case (as we do READ_UNCOMMITTED).

So this patch works on removing this transaction support altogether for the
histogram sampling code.

RB: 25015
Reviewed-by: Jakub Lopuszanski <jakub.lopuszanski@oracle.com>