Bug #30660403 ROLES NOT HANDLING COLUMN LEVEL PRIVILEGES CORRECTLY; CAN SELECT, BUT NOT UPDATE
Description:
When assigning column level privileges using Roles, a schema level
SELECT privilege is not being considered when making a column level
UPDATE statement with a WHERE clause such that the UPDATE statement
will fail with the error:

Error Code: 1143. SELECT command denied to user
'TestUserFails'@'localhost'
for column 'test_id' in table 'test_table'

However, this works as expected for Users, so it seems to be a bug
with how Roles are handled.

Analysis:
For the update query the in case of non active roles the
grant->privilege is updated in the check_access() with the db access
priv and then in check_grant(),
the privileges are appended with the grant_tables. And in
check_grant_column()
the updated grant->privilege are verified.
But in case of active roles the privileges are over written with the
aggr.table_access
becasue of that in check_grant_column() the grant->privilege is '0'
i.e select privilege is absent. Hence we see the error

Fix: Append the aggr.table_access to the grant.privilege instead of
over writting, similarly we do it in case of non active roles.

RB: 25004