Bug #31818486: FIX MSVC C4319 WARNINGS
When building tests lot0lot-t, ut0frags-t, ut0plist-t and zlob0int-t as
64 bit executables, the MSVC warning C4319 "'~' : zero extending 'type1'
to 'type2' of greater size" is generated on Windows.

These warnings are generated when building 64 bit applications on
Windows because the ulint type is defined as unsigned long for the offending
unit tests. The unsigned long is 32 bits on Windows, and combining 32 bit
ulint values in boolean bitwise expressions with 64 bit pointers generates the
C4319 warning.  The ulint type is defined as unsigned __int64 elsewhere
in the MySQL source, and defining ulint as unsigned __int64 in the
offending unit tests for 64 bit Windows fixes the generation of the C4319
warning.

Reviewed by: Annamalai Gurusami <annamalai.gurusami@oracle.com>