Bug#31750214: RECENT CRASH IN QEP_SHARED_OWNER::TABLE
Problem:
This problem is a regression from WL#8084.
If a semi-join condition that is part of a WHERE clause is pushed
down to a materialized derived table, the columns that are part of
the condition are replaced with expressions from the derived table.
However, the same columns are part of sj_inner_exprs/sj_outer_exprs.
As a result, the replacement in the condition also results in
replacement in sj_inner_exprs/sj_outer_exprs. This leads to a mismatch
of table reference in these lists. The table reference for the
expressions in sj_inner_exprs/sj_outer_exprs need to be that of the
derived table while the expressions of the pushed down condition
has to be that of the derived tables's underlying table.

Solution:
Before replacing the columns in a condition(being pushed down) with
derived table expressions, optimizer now checks if the derived table
is part of a semi-join. If yes, it checks if the expressions
in the condition are part of semi-join expressions, both inner and
outer. If there is a match, expressions from the sj_inner_exprs and
sj_outer_exprs are removed.

Reason for test result changes:
If the removal of expressions from the semi-join expression lists
results in an empty list, we add a const condition "1=1" similar
to what is done in build_sj_cond(). As a result, for the query that
sees the plan change, condition "7=table1.col_int_key" is pushed down,
so it's now changed to "1=1". This makes the materialized subquery const
at execution. It is read only once.

Change-Id: I57bd7e360e1f153bc6113513148255dd77c56e9d