Bug#31749103: LOG TIMESTAMP "TS" ADDED BY WL#13786 DOES NOT AGREE WITH OTHER TIMESTAMPS
The timestamps with the "ts" key which were added by
WL#13786 were reported as being several minutes in the
future, and as not agreeing with other log timestamps
(e.g. the ISO8601 timestamp, "time").

In my_micro_time(), we use

  return (static_cast<unsigned long long int>
  (t.tv_sec) * 1000000 + t.tv_usec);

As we wish to combine seconds and microseconds into one
longlong, we shift the seconds value to the left by six
places, and then add the six places of the microsecond
value in the "hole" we made.

  return (static_cast<unsigned long long int>
  (t.tv_sec) * 1000 + t.tv_usec);

In my_milli_time(), we want milliseconds, so we shift
the seconds to the left by three places. However, we
then add the six places of the microsecond value into
the "hole" three wide. Since the microsecond value is
off by three orders of magnitudes, its leftmost three
places affect seconds and event minutes of tv_sec,
enough to put the combined value several minutes "into
the future."

This patch hotfixes the assignment to read

  return (static_cast<unsigned long long int>
  (t.tv_sec) * 1000 + (t.tv_usec/1000));

while we evaluate whether this key/value pair is still
needed.

Approved by: Jens Even Blomsoy <jens.even.blomsoy@oracle.com>