Bug #31739513 SEPARATE JSON (DE)SERIALIZATION CODE FROM MYSQL SPECIFIC CODE
This patch creates a standalone shared library 'json_binlog' to
de-serialize json data.  The public interface is in sql/json_binary.h
which provides functions for parsing binary data, and converting it to
Value objects. We also provide utility functions to convert Value
objects to JSON strings.

The source file sql/json_binary.h has been cleaned up by removing
misc. MySQL'isms like DBUG_ASSERT and integer types such as 'uint8'
and 'uint32'.  Functions that are only available in server mode have
been wrapped in ifdef MYSQL_SERVER.

There is a new cmake option WITH_JSON_BINLOG_LIBRARY which defaults to
OFF.  If this is ON, we will link the mysqlbinlog client with the
shared library.  If the option is ON, we will also INSTALL the
library, and the accompanying header file.

The WITH_JSON_BINLOG_LIBRARY option is currently not switched on for
any of our regular builds. It is intended for MySQL users who need to
de-serialize JSON data, and who do their own builds.

Change-Id: Iac434cda7b072dcd2d6c3dcd946a1da8fb3a423f