Bug#31769242 DUPLICATE WARNING WITH PS-PROTOCOL FOR CREATE TABLE ... SELECT READING ACL TABLE
Post-push fix for wl#14087.

The problem was that we get duplicate warnings when executing SQL statement
which reads ACL tables in serializable or repeatable read mode.

This patch avoids requesting SE to skip row locks for ACL tables during
prepare phase. It is OK to do so as during this phase there will be no
reads from SE anyway. We still request SE to skip row locks for ACL tables
during execution phase.

Change-Id: I44ea4e7a91794c7807772d0ee848eda5113a3640