Bug#31836032 timer with sub-millisec waste CPU
Problem
=======

net::timer can specify a duration it should wait before it expires.

The resolution of that timer depends on the clock that's used, but quite
likely is in nanoseconds.

The wait is implemented via poll()/epoll()/... which have millisecond
resolution.

If the expiry is less than a millisecond, e.g. 900us, poll() would
convert it to 0 milliseconds and not wait.

Instead the check that checks if the timer is expired would enter a
tight loop until the 900us are passed and then leave.

Change
======

- round the expiry to next full millisecond before passing it to
poll()/epoll()/... to wait /at least/ the expected time.

RB: 25063