Bug#31668306 ALLOW DYNAMICALLY LINKED NDB TOOLS
Add a CMake option NDB_UTILS_LINK_DYNAMIC to control whether tools
such as ndb_drop_table are linked with ndbclient using static or
dynamic linking.  This option defaults to OFF.

This patch uncouples NDBTEST from NDBCLIENT in the CMake function
ndb_add_executable(). The majority of the lines of code in the
patch are actually concerned with this uncoupling.

Reviewed-by: Tor Didriksen <tor.didriksen@oracle.com>