Bug#31589019 Using optimal memory-barrier (in form of acquire/release) for event-mutex
This patch is based on contribution by Krunal Bauskar from Huawei. It:

1. Removes the `std::atomic_thread_fence(std::memory_order_acquire);` call, as
it was not helping in the proof of correctness, translated to a no-op on x86,
and to a costly dmb ishld barrier on ARM64.
2. Changes all the operations on `m_waiters` to memory_order_seq_cst. The cost
of this change is that on Intel we have one additional mfence or xchg on the
slow path leading to sleep, and another on the slow path leading to wake up -
hopefully both should be rare. Anyway, correctness triumphs performance, and I
couldn't prove correctness in the light of C++14 standard without this.
3. Provides a rigorous proof of the theorem that a thread which calls
`sync_array_wait_event` will not end up sleeping indefinitely.
4. On ARM compiles to the same assembly as the code initially proposed by
Krunal Bauskar, see https://godbolt.org/z/e1cha3 for comparison.
5. On Intel compiles to assembly which does not have the bug present in the
original patch provided by contributor, which let the CPU flush the store to
m_lock_word after reading m_waiters.
6. Removes the unused `waiters()` method.

Reviewed-by: Nikša Skeledžija <niksa.skeledzija@oracle.com>
Reviewed-by: Marcin Babij <marcin.babij@oracle.com>
RB:24804