Bug#31844042 PB2: NDB.NDB_ERROR_REPORTER FAIL WITH: NODEGROUP 0 HAS 1  MEMBERS, NOOFREPLICAS=2
Problem
=======

Due to Bug#31825181 NODEGROUP WAS NOT ASSIGNED AFTER FIRST USE OF
NODEGROUP=65536. DBDIH LINE: 2490 which caused nodegroup checking in
ndb_mgmd to stop at first occurrence of Nodegroup=65536 the config in
ndb_error_reporter.ini passed when it should fail.

Bug#31825181 was thought to been introduced in 8.0.18 with new config,
but turns out that it was an ancient bug, even being in 7.2.

Now when that bug is fixed ndb.ndb_error_reported started to fail with:

[ndb_config] ERROR    -- at line 26: Nodegroup 0 has 1 members, NoOfReplicas=2

Solution
========

Add node definitions such that number of nodes in nodegroup 0 is two.

Reviewed-by: Priyanka Sangam <priyanka.sangam@oracle.com>