Bug#31781948 UPDATE NODE.JS TO 14.4.0, 12.18.0 OR 10.21.0
Remove storage/ndb/nodejs from Cluster 7.x.

Recent versions of Node.JS including Node.js 12.18, which is
now used to build MySQL Cluster 8.0, are not supported by the
toolchain used to build MySQL Cluster 7.x on most platforms.

Remove all of storage/ndb/nodejs, the text in the top-level
LICENSE file that pertains to it, and any code under packaging/
that refers to it.

Reviewed-by: Lakshmi Narayanan Sreethar <lakshmi.narayanan.sreethar@oracle.com>