Fix sporadic test failures under ASAN, related to lost memory in fil_tablespace_redo_encryption() when the function is left without adding or updating a key.
Key storage can be allocated before the decision is made, whether a key
needs to be added/updated at all. If this decision evaluates to false,
the function is left without freeing the allocated storage.

The storage is also not freed if decoding of the encryption info fails.

Fix by moving the above mentioned decision to the top of the function,
so that the storage is allocated only when needed.

Also free the storage, if the decoding fails.

RB#24962