Bug #31397578: MAKE TESTS RESILIENT TO ORDER CHANGES [noclose]
Yet more test changes to prepare for the hypergraph optimizer.

The biggest change this time around is that many optimizer trace
tests are now explicitly skipped, instead of relying on the hypergraph
optimizer to throw an error itself -- this is because we want to be able
to query the optimizer trace even when the hypergraph optimizer is active,
even if it does not return the same data.

Change-Id: I7b1cfd706c9364d8e904953e97287e5006dacd5c