Bug#31883796 NDB MTR TESTS IN VALGRIND OFTEN FAIL DUE TO A DEADLOCK
MTR executes a test called check-warnings.test after running every test
to check for any unexpected warnings found in the server log. This test
is run in parallel against every running MySQL Server to save time. The
check-warnings.test involves executing a procedure called
mtr.check_warnings. This procedure consists of a DDL in its definition
causing any ndbcluster enabled MySQL Server to acquire a GSL when it is
executed. In normal builds this should be a very quick acquire and
release and even when running in parallel it will not cause any issues.
But in the valgrind builds, this is slower than the normal, and, often
during check warnings, multiple runs timeout waiting for the GSL and end
up failing the test itself.

Fixed that issue in this patch by running check-warnings.test one by one
for all the MySQL Servers if they are running with ndbcluster and
valgrind.

Change-Id: Id1f0f5d5037a8d839d180a454ab9dab5b37e8efa