Bug#31327337: SUBSYSTEM NAME FOR SEMISYNC ERRORS IS [SERVER]
Messages from semi-sync replication appeared in the error-log
with a subsystem tag of "Server", rather than the expected
"Repl". This was a result of a fault in the per-subtree set-up
of semi-sync.

Messages from semi-sync, such as MY-011171
(ER_SEMISYNC_STOP_BINLOG_DUMP_TO_SLAVE) are now correctly
issued with a subsystem tag of "Repl".

This incidently is a great example of how easy testing for
error-log contents is since the addition of the table
performance_schema.error_log (WL#13681):

> SELECT subsystem, error_code
    FROM performance_schema.error_log
   WHERE error_code="MY-011170" LIMIT 1;

subsystem      error_code
Repl   MY-011170

Change-Id: I5ae349338c46d75d7e7d7641cd8bff6404b95933