Bug#31819558 SHOW VARIABLES LIKE 'PARTIAL_REVOKES' REPORTS INCORRECT OUTPUT AFTER THE RESTART
Description:
------------

Server restart with --partial_revokes=OFF is ignored with a warning in log file
as long as there exists at least one restriction in the  mysql.user table. It
means that after the restart, the SHOW VARIABLES stmt must report the variable
value to be ON but it may report the value to be OFF.

This is the problem only with reporting the status variable. Internally partial
revokes is tracked through an atomic variable 'partial_revokes' that gets
updated correctly i.e. ON.

The reason variable value is reported incorrectly because SHOW VARIABLES fetches
the value from the performance_schema's global_variables table.
Performance_schema uses the sys_var structures to track the value.
We pass a global boolean variable 'opt_partial_revokes' to the
Sys_partial_revokes structure because it does not support the atomic variables
yet. During the server restart Sys_partial_revokes structure's check() and
update() methods are not invoked. Hence the global variable is not set to the
expected state.

Fix
---
1. During server restart if found at least one restriction in the user table
then  set the 'opt_partial_revokes' to true as well.  Added a mutex to  update
the global variable safely.

2. During review I found another issue related to FLUSH PRIVILEGES. Introduced a
mutex to protect the global variable while upating it. Atomic variable is also
updated togather since it is mostly used to read operations throughout the ACL
code.

Testing
-------

1. Added SHOW VARIABLES stmts in the existing test files.
2. Corrected an existing test case. It was not complete.
3. Added test for FLUSH PRIVILEGES and SET PERSIST stmts.
4. verified on loki.

Review
-------
RB#25062