Bug#30838749: PERFORMANCE DROP WHEN CREATING A FUNCTIONAL INDEX
After adding a functional index or a multi-valued index to a table, a
query that performed SELECT * on that table, without actually using
the index, would be slower than it was if the index was not present.

The reason for the slowdown was that the value of the hidden generated
column backing the index was calculated for each row. This was
unnecessary work, since the value of the hidden column was not used by
the query.

Calculation of the hidden column happened as a consequence of the fix
for bug#28069731. The problem in that bug was that HANDLER statements
added the hidden column to the table's read set, but not to its write
set, so that an assertion failure was raised when trying to update the
hidden column. The bug was fixed by adding the hidden column to the
write set. This was done in code shared between HANDLER statements and
SELECT * statements (in insert_fields()), and it had the side-effect
of also adding the hidden column to the read set and the write set in
SELECT * statements.

This patch reworks the fix for bug#28069731 so that instead of adding
the hidden column to the write set for both HANDLER statements and
SELECT * statements, it removes the hidden column from the read set in
HANDLER statements. This way, neither HANDLER statements nor SELECT *
statements calculate the value of hidden columns that are not going to
be used in the query. The hidden column is still added to the read set
and write set of the table if the optimizer determines that the index
backed by the column should be used.

Change-Id: I51576c28f48eeb64971a253c875d36a874fc338b