Bug #19186271 MANY, MANY CLASSES OF STRING-TO-INTEGER CONVERSION BUG [two]
Four distinct classes of string-to-integer conversion bug.
This patch fixes one of the issues:

(2) Decimals of the form '0.9...' with more than 19 digits, which happen to
lie between 0.9223372036854775807 (this is 0.<LLONG_MAX>) and 1.0 exclusive,
are "rounded downward" to 0 rather than rounded to 1.

We had this code for determining whether to round up:
  if (r >= d) ull++;
The problem was that the code for calculating 'r' could overflow to zero.

So instead we test for:
  if (r/2 >= d/2)
which is equivalent, since d is always a multiple of 10 (and hence a multiple of 2).
This avoids the multiplication by two here: r = (ull % d) * 2;

Change-Id: I98c3d14d93d0728f1021cf71c8e307d16cfd5b09
(cherry picked from commit e3e114f8e344f5e28fb7e0015ffa0987dfb7f227)