REGRESSION IN TEST MEB.COMPRESSION WITH Bug#31396947 FIX
Issue
=====
os_file_compress_page() requires that the page size is a multiple of file
system block size. It has an assert which checks this condition.  During
import, while checking if a page needs to compressed or not, we do not check
for this condition, resulting in the assert being hit on windows platform
where block size is greater than page size.

Fix
===
Do compression only if page size is a multiple of file system block size
(provided that other requirements for compression are satisfied).

RB: 25106
Reviewed-by: Mateusz Rukowicz <mateusz.rukowicz@oracle.com>