Bug#31377118: SERVER START CRASHES WHEN USING NON EXISTING TMPDIR
Background

Change-Id: I0192a22ebee183140c409b15dc223fe882049a1a
----------
When starting the server, it will use one or more temporary
directories for various purposes. The list of temporary directories
is constructed by the following actions (ordered by priority) during
server startup:

1. Submitting a list of paths as a startup or config option.
2. Setting an environment variable.
3. Relying on the platform default.

If a list of directories is submitted, the directories will be used
in a round robin fashion to spread the files over various locations.

Problem
-------
The problem was that when a temporary file was created, the server
did not check for the outcome of the mkstmp() system call. If this
failed, and an invalid file descriptor was returned, the code still
carried on.

Solution
--------
This patch introduces two checks to improve the handling of this:

1. During server start, verify that each temporary directory exists.
   This is not strictly necessary (see below), but it makes sense to
   catch issues like this as early as possible.
2. When creating a file in the tmpdir, check that the file is indeed
   successfully created, and otherwise, report an error. This is
   necessary even if the check during server start (see above) is in
   place because a directory may be deleted after the server has
   been started.

Some refactoring had to be done while implementing the second item
above in order to handle the error in a graceful way instead of just
crashing a bit later.

There is also a new MTR test case to verify that the fix works as
expected.

Change-Id: I7592a64a51586cbd32a95175294610e409dcc23a