Bug#31667405: Incorrect prepared statement outcome with numeric strings in JSON
WL#9384 implemented consistent behaviour of JSON parameters with regard
to how the parameter was used. Before, JSON parameters used in e.g.
INSERT and UPDATE, and where a character string value was supplied,
were treated as JSON strings. This happened regardless of whether the
parameter was wrapped in a CAST(AS JSON) clause. But when the parameter
was used in a comparison operator, e.g. when compared to the result of
JSON_EXTRACT, it was treated as a scalar value. WL#9384 changed this
by always treating the parameter as a JSON object.

However, this has been deemed incompatible with prior versions and
behavior will be reverted. The fix is implemented by adding a flag to
the Item_param class (m_json_as_scalar) that tells how to handle
character strings that are processed by Item_param::val_json():
- As default, such parameters are interpreted as JSON objects.
- Within a comparison operator and as an argument to the IN operator,
  such parameters are attempted to be interpreted as scalar values, unless
- When wrapped in a CAST AS JSON clause, they are interpreted as
  JSON objects.

Also fixed another problem: When failing to identify a JSON string,
the server might accept the string as a JSON scalar. Now this produces
an error message instead.

Reviewed by: Knut Anders Hatlen <knut.hatlen@oracle.com>